chore: improve strict Clippy compliance and strengthen CI checks
- add explicit Clippy allow attributes for intentional numeric casts
- document intentional float/integer conversion behavior
- improve handling of strict pedantic/nursery lint configuration
- upgrade CI Clippy step to use:
  --all-targets
  --all-features
  clippy::all
  clippy::pedantic
  clippy::nursery
- add documentation comments explaining strict lint policy
- strengthen automated code quality enforcement in GitHub Actions

+++ pasja_informatyki/odc-5/simple-example-srednia/src/main.rs
@@ -15,8 +15,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         *ocena = wejście.trim().parse()?;
     }
 
-    let suma = oceny.iter().sum::<f64>().round();
-    let ilość_ocen = oceny.iter().len() as f64;
+    let suma = oceny.iter().sum::<f32>().round();
+    #[allow(clippy::cast_precision_loss)]
+    let ilość_ocen = oceny.iter().len() as f32;
     let średnia_arytmetyczna = suma / ilość_ocen;
 
     println!("oceny: {oceny:?}");
